Create articles comp and move logic out of app.js
