+++ src/App.js
@@ -1,23 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux'
 
-import { fetchArticlesBySub } from '../src/features/article/articleSlice'
-import Article from '../src/features/article/Article'
+import Articles from './features/article/Articles'
 
 function App() {
 
   const [subreddit, setSubreddit] = useState('webdev')
   const [subFound, setSubFound] = useState(true)  
-  
-  const dispatch = useDispatch()
-
-  useEffect(() => {
-    dispatch(fetchArticlesBySub(subreddit))
-  }, [subreddit])
-
-  const articles = useSelector((state) => state.article.articles)
-
-
 
   return (
     <div className="App">
@@ -25,11 +13,7 @@ function App() {
         <input type="text" name="sub" id="sub" className="input" value={subreddit} onChange={e => setSubreddit(e.target.value)}/>
         <label htmlFor="sub" hidden={subFound}>No subs found!</label>
       </header>
-      <div className="articles">
-        {
-          (articles) ? articles.map(article => <Article key={article.data.id} article={article.data} />) : ''
-        }
-      </div>
+      <Articles subreddit={subreddit} />
     </div>
   );
 }
